requirements.txt
torch 
tensorflow
pandas
dvc
notebook
matplotlib
numpy
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
scipy
Flask
Flask-cors
fastapi
kaggle
python-dotenv
-e .